Change BQM convert functions to be class methods
