+++ dimod/binary_quadratic_model.py
@@ -702,7 +702,7 @@ class BinaryQuadraticModel(object):
         Examples:
             This example creates an Ising model with a positive energy offset, and
             then removes it.
-             
+
             >>> bqm = dimod.BinaryQuadraticModel({}, {}, 1.3, dimod.SPIN)
             >>> bqm.remove_offset()
             >>> bqm.offset
@@ -880,7 +880,7 @@ class BinaryQuadraticModel(object):
 
         Args:
             bqm (:class:`.BinaryQuadraticModel`):
-                The updating binary quadratic model. Any variables in the updating 
+                The updating binary quadratic model. Any variables in the updating
                 model are added to the updated model. Values of biases and the offset
                 in the updating model are added to the corresponding values in
                 the updated model.
@@ -1242,3 +1242,337 @@ class BinaryQuadraticModel(object):
         en += sum(linear[v] * sample[v] for v in linear)
         en += sum(sample[u] * sample[v] * quadratic[(u, v)] for u, v in quadratic)
         return en
+
+##################################################################################################
+# conversions
+##################################################################################################
+
+    def to_networkx_graph(self, node_attribute_name='bias', edge_attribute_name='bias'):
+        """Return the BinaryQuadraticModel as a NetworkX graph.
+
+        Args:
+            node_attribute_name (hashable):
+                The attribute name for the linear biases.
+            edge_attribute_name (hashable):
+                The attribute name for the quadratic biases.
+
+        Returns:
+            :class:`networkx.Graph`: A NetworkX graph with the biases stored as
+            node/edge attributes.
+
+        Examples:
+            >>> import networkx as nx
+            >>> bqm = dimod.BinaryQuadraticModel({0: 1, 1: -1, 2: .5},
+            ...                                  {(0, 1): .5, (1, 2): 1.5},
+            ...                                  1.4,
+            ...                                  dimod.SPIN)
+            >>> BQM = bqm.to_networkx_graph()
+            >>> BQM[0][1]['bias']
+            0.5
+            >>> BQM.node[0]['bias']
+            1
+
+            Also, if the preferred notation is 'weights'
+
+            >>> import networkx as nx
+            >>> bqm = dimod.BinaryQuadraticModel({0: 1, 1: -1, 2: .5},
+            ...                                  {(0, 1): .5, (1, 2): 1.5},
+            ...                                  1.4,
+            ...                                  dimod.SPIN)
+            >>> BQM = bqm.to_networkx_graph(edge_attribute_name='weight')
+            >>> BQM[0][1]['weight']
+            0.5
+
+
+        """
+        import networkx as nx
+
+        BQM = nx.Graph()
+
+        # add the linear biases
+        BQM.add_nodes_from(((v, {node_attribute_name: bias, 'vartype': self.vartype})
+                            for v, bias in iteritems(self.linear)))
+
+        # add the quadratic biases
+        BQM.add_edges_from(((u, v, {edge_attribute_name: bias}) for (u, v), bias in iteritems(self.quadratic)))
+
+        # set the offset and vartype properties for the graph
+        BQM.offset = self.offset
+        BQM.vartype = self.vartype
+
+        return BQM
+
+    def to_ising(self):
+        """Converts the binary quadratic model into the (h, J, offset) Ising format.
+
+        If the binary quadratic model's vartype is not spin, it is converted.
+
+        Returns:
+            tuple: A 3-tuple:
+
+                dict: The linear biases.
+
+                dict: The quadratic biases.
+
+                number: The offset.
+
+        """
+        return self.spin.linear, self.spin.quadratic, self.spin.offset
+
+    @classmethod
+    def from_ising(cls, h, J, offset=0.0):
+        """Build a binary quadratic model from an Ising problem.
+
+
+        Args:
+            h (dict[variable, bias]/list[bias]):
+                The linear biases of the Ising problem. If a list, the indices of the list are treated
+                as the variable labels.
+
+            J (dict[(variable, variable), bias]):
+                The quadratic biases of the Ising problem.
+
+            offset (optional, default=0.0):
+                The constant offset applied to the model.
+
+        Returns:
+            :class:`.BinaryQuadraticModel`
+
+        """
+        if isinstance(h, list):
+            h = dict(enumerate(h))
+
+        return cls(h, J, offset, Vartype.SPIN)
+
+    def to_qubo(self):
+        """Converts the binary quadratic model into the (Q, offset) QUBO format.
+
+        If the binary quadratic model's vartype is not binary, it is converted.
+
+        Returns:
+            tuple: A 2-tuple:
+
+                dict: The qubo biases. A dict where the keys are pairs of variables and the values
+                are the associated linear or quadratic bias.
+
+                number: The offset.
+
+        """
+        qubo = {}
+
+        for v, bias in iteritems(self.binary.linear):
+            qubo[(v, v)] = bias
+
+        for edge, bias in iteritems(self.binary.quadratic):
+            qubo[edge] = bias
+
+        return qubo, self.binary.offset
+
+    @classmethod
+    def from_qubo(cls, Q, offset=0.0):
+        """Build a binary quadratic model from a qubo.
+
+        Args:
+            Q (dict):
+                The qubo coefficients.
+
+            offset (optional, default=0.0):
+                The constant offset applied to the model.
+
+        Returns:
+            :class:`.BinaryQuadraticModel`
+
+        """
+        linear = {}
+        quadratic = {}
+        for (u, v), bias in iteritems(Q):
+            if u == v:
+                linear[u] = bias
+            else:
+                quadratic[(u, v)] = bias
+
+        return cls(linear, quadratic, offset, Vartype.BINARY)
+
+    def to_numpy_matrix(self, variable_order=None):
+        """Return the binary quadratic model as a matrix.
+
+        Args:
+            variable_order (list, optional):
+                If variable_order is provided, the rows/columns of the numpy array are indexed by
+                the variables in variable_order. If any variables are included in variable_order that
+                are not in `self`, they will be included in the matrix.
+
+        Returns:
+            :class:`numpy.matrix`: The binary quadratic model as a matrix. The matrix has binary
+            vartype.
+
+        Notes:
+            The matrix representation of a binary quadratic model only makes sense for binary models.
+            For a binary sample x, the energy of the model is given by:
+
+            .. math::
+
+                E(x) = x^T Q x
+
+
+            The offset is dropped when converting to a numpy matrix.
+
+        """
+        import numpy as np
+
+        if variable_order is None:
+            # just use the existing variable labels, assuming that they are [0, N)
+            num_variables = len(self)
+            mat = np.zeros((num_variables, num_variables), dtype=float)
+
+            try:
+                for v, bias in iteritems(self.binary.linear):
+                    mat[v, v] = bias
+            except IndexError:
+                raise ValueError(("if 'variable_order' is not provided, binary quadratic model must be "
+                                  "index labeled [0, ..., N-1]"))
+
+            for (u, v), bias in iteritems(self.binary.quadratic):
+                if u < v:
+                    mat[u, v] = bias
+                else:
+                    mat[v, u] = bias
+
+        else:
+            num_variables = len(variable_order)
+            idx = {v: i for i, v in enumerate(variable_order)}
+
+            mat = np.zeros((num_variables, num_variables), dtype=float)
+
+            try:
+                for v, bias in iteritems(self.binary.linear):
+                    mat[idx[v], idx[v]] = bias
+            except KeyError as e:
+                raise ValueError(("variable {} is missing from variable_order".format(e)))
+
+            for (u, v), bias in iteritems(self.binary.quadratic):
+                iu, iv = idx[u], idx[v]
+                if iu < iv:
+                    mat[iu, iv] = bias
+                else:
+                    mat[iv, iu] = bias
+
+        return np.asmatrix(mat)
+
+    @classmethod
+    def from_numpy_matrix(cls, mat, variable_order=None, offset=0.0, interactions=None):
+        """Build a binary quadratic model from a numpy matrix.
+
+        Args:
+            mat (:class:`numpy.matrix`):
+                A square numpy matrix. The coefficients of a qubo.
+
+            variable_order (list, optional):
+                If variable_order is provided, provides the labels for the variables in the binary
+                quadratic program, otherwise the row/column indices will be used. If variable_order
+                is longer than the matrix, the extra values are ignored.
+
+            offset (optional, default=0.0):
+                The constant offset for the binary quadratic program.
+
+            interactions (iterable, optional, default=[]):
+                Any additional 0.0-bias interactions to be added to the binary quadratic model.
+
+        Returns:
+            :class:`.BinaryQuadraticModel`
+
+        """
+        import numpy as np
+
+        if mat.ndim != 2:
+            raise ValueError("expected input mat to be a square matrix")  # pragma: no cover
+
+        num_row, num_col = mat.shape
+        if num_col != num_row:
+            raise ValueError("expected input mat to be a square matrix")  # pragma: no cover
+
+        if variable_order is None:
+            variable_order = list(range(num_row))
+
+        if interactions is None:
+            interactions = []
+
+        bqm = cls({}, {}, offset, Vartype.BINARY)
+
+        for (row, col), bias in np.ndenumerate(mat):
+            if row == col:
+                bqm.add_variable(variable_order[row], bias)
+            elif bias:
+                bqm.add_interaction(variable_order[row], variable_order[col], bias)
+
+        for u, v in interactions:
+            bqm.add_interaction(u, v, 0.0)
+
+        return bqm
+
+    def to_pandas_dataframe(self):
+        """Return the binary quadratic model as a pandas DataFrame.
+
+        Returns:
+            :class:`pandas.DataFrame`: The binary quadratic model as a DataFrame. The DataFrame has
+            binary vartype. The rows and columns are labeled by the variables in the binary quadratic
+            model.
+
+        Notes:
+            The DataFrame representation of a binary quadratic model only makes sense for binary models.
+            For a binary sample x, the energy of the model is given by:
+
+            .. math::
+
+                E(x) = x^T Q x
+
+
+            The offset is dropped when converting to a pandas DataFrame.
+
+        """
+        import pandas as pd
+
+        try:
+            variable_order = sorted(self.linear)
+        except TypeError:
+            variable_order = list(self.linear)
+
+        return pd.DataFrame(self.to_numpy_matrix(variable_order=variable_order),
+                            index=variable_order,
+                            columns=variable_order)  # let it choose its own datatype
+
+    @classmethod
+    def from_pandas_dataframe(cls, bqm_df, offset=0.0, interactions=None):
+        """Build a binary quadratic model from a pandas dataframe.
+
+        Args:
+            bqm_df (:class:`pandas.DataFrame`):
+                A pandas dataframe. The row and column indices should be that variables of the binary
+                quadratic program. The values should be the coefficients of a qubo.
+
+            offset (optional, default=0.0):
+                The constant offset for the binary quadratic program.
+
+            interactions (iterable, optional, default=[]):
+                Any additional 0.0-bias interactions to be added to the binary quadratic model.
+
+        Returns:
+            :class:`.BinaryQuadraticModel`
+
+        """
+        if interactions is None:
+            interactions = []
+
+        bqm = cls({}, {}, offset, Vartype.BINARY)
+
+        for u, row in bqm_df.iterrows():
+            for v, bias in row.iteritems():
+                if u == v:
+                    bqm.add_variable(u, bias)
+                elif bias:
+                    bqm.add_interaction(u, v, bias)
+
+        for u, v in interactions:
+            bqm.add_interaction(u, v, 0.0)
+
+        return bqm
